Adding entries for E&O OD 2025 beta release
Attempting an addition of the information for the 2025 beta release of
Education and Outreach Open Data - 13 TeV pp data. Much of this relies
on the identicality between the datasets released for research and those
released for E&O, which is nice to be able to take advantage of.

We have several skims of these data available, so the skims are packed
into different "releases" here - which seemed the most convenient way to
represent these data without inventing a whole new scheme of requesting
samples and confusing users about what they're doing.