Adding direct eos access support
Adding direct eos access support

For those who have eos fuse-mounted, this allows them to use 'eos' as a protocol and simply get eos paths. There should be no issue, in principle, with using xrootd, but I am imagining that there might be some system (not sure if this is possible?) that doesn't have xrootd installed but does have eos installed. If somebody knows for sure that's not possible, then we shouldn't bother with this.