Adding type hints, dropping python 3.9
This addresses #38

Adding type hints to all the functions in metadata and utils

In Python 3.9, we can't use the `str | None` style specification for
optional arguments; we instead have to use `Optional`. It feels to me
like it is not worth cutting up our code base to retain support for
python 3.9. On the other hand (annoyingly), lxplus still uses python 3.9
by default, so this means folks need a non-standard version of python
out of the box to run on lxplus. I'm not sure what's best here, but this
is a version submitted for your consideration.

In the meantime, this PR will also complete our test coverage.